# perf(check): use v8 code cache for extension sources in `deno check`